requirements.txt
streamlit>=1.28.0
openai>=1.3.0
faiss-cpu>=1.7.4
pymupdf>=1.23.0
python-pptx>=0.6.23
python-docx>=0.8.11
pandas>=2.0.0
sentence-transformers>=2.2.0
pydantic>=2.0.0
numpy>=1.21.0
asyncio-dispatch>=1.0.0
